Updated Project
Added the following:

As the user plays, the current high score will be displayed on the screen alongside the player name who created the score.
If the high score is beaten, the new score and player name will be displayed instead.
The highest score will be saved between sessions, so that if the player closes and reopens the application, the high score and player name will be retained.